add ml training wrappers
